small test cleanup
